3/31

+++ src/main/java/knou/lms/mrk/vo/MarkSubjectVO.java
@@ -15,9 +15,9 @@ public class MarkSubjectVO extends DefaultVO {
     private String profMemo;        // 교수 메모
     private String drvtnMrkGrdcd;   // 산출성적 등급코드
     private String passyn;          // 통과여부
-    private int totScr;             // 총점 (최종점수 + 가산점수)
-    private int lstScr;             // 최종 점수 (총점 - 가산점수)
-    private int adtnScr;            // 가산 점수 (= 기타점수)
+    private double totScr;             // 총점 (최종점수 + 가산점수)
+    private double lstScr;             // 최종 점수 (총점 - 가산점수)
+    private double adtnScr;            // 가산 점수 (= 기타점수)
 
 
     public String getMrkSbjctId() {
@@ -68,27 +68,27 @@ public class MarkSubjectVO extends DefaultVO {
         this.passyn = passyn;
     }
 
-    public int getTotScr() {
+    public double getTotScr() {
         return totScr;
     }
 
-    public void setTotScr(int totScr) {
+    public void setTotScr(double totScr) {
         this.totScr = totScr;
     }
 
-    public int getLstScr() {
+    public double getLstScr() {
         return lstScr;
     }
 
-    public void setLstScr(int lstScr) {
+    public void setLstScr(double lstScr) {
         this.lstScr = lstScr;
     }
 
-    public int getAdtnScr() {
+    public double getAdtnScr() {
         return adtnScr;
     }
 
-    public void setAdtnScr(int adtnScr) {
+    public void setAdtnScr(double adtnScr) {
         this.adtnScr = adtnScr;
     }
 }
